+++ test/instance.test.js
@@ -1,5 +1,6 @@
 import HttpRequestMock from '../src/index';
 import FetchInterceptor from '../src/interceptor/fetch';
+import MockItem from '../src/mocker/mock-item';
 import Mocker from '../src/mocker/mocker';
 
 describe('test HttpRequestMock necessary methods', () => {
@@ -28,6 +29,14 @@ describe('test HttpRequestMock necessary methods', () => {
     expect(HttpRequestMock.disable()).toBe(new Mocker());
   });
 
+  it('enableLog method should return an instance of Mocker', () => {
+    expect(HttpRequestMock.enableLog()).toBe(new Mocker());
+  });
+
+  it('disableLog method should return an instance of Mocker', () => {
+    expect(HttpRequestMock.disableLog()).toBe(new Mocker());
+  });
+
   it('Mocker.setMockData method should return an instance of Mocker', () => {
     expect(new Mocker().setMockData({'key': {}})).toBe(new Mocker());
   });
@@ -49,7 +58,7 @@ describe('test HttpRequestMock necessary methods', () => {
     mocker.enable();
     expect(mocker.matchMockItem('http://www.api.com/mock')).toBe(null);
 
-    mocker.mock({url: 'http://www.api.com/regexp-mock', regexp: ['\/regexp-mock$', '']});
+    mocker.mock({url: new RegExp('\/regexp-mock$')});
     expect(mocker.matchMockItem('http://www.api.com/regexp-mock')).toBeTruthy();
   });
 
@@ -65,6 +74,42 @@ describe('test HttpRequestMock necessary methods', () => {
     expect(info3.body).toMatchObject({ rst: 789 });
 
     expect(instance).toBe(new FetchInterceptor(mocker));
+  });
+
+  it('Mocker.disableLog method should return an instance of Mocker', () => {
+    expect(new Mocker().disableLog()).toBe(new Mocker());
+  });
+
+  it('Mocker.enableLog method should return an instance of Mocker', () => {
+    const [log1, log2, log3] = [console.groupCollapsed, console.groupEnd, console.log];
+    const mocker = new Mocker();
+    console.groupCollapsed = jest.fn();
+    console.groupEnd = jest.fn();
+    console.log = jest.fn();
+    expect(mocker.enableLog()).toBeInstanceOf(Mocker);
+    mocker.groupLog([123, 'abc']);
+    mocker.groupLog([[123, 456], ['abc', 'xyz']]);
+    expect(console.groupCollapsed).toBeCalled();
+    expect(console.groupEnd).toBeCalled();
+
+    console.groupEnd = undefined;
+    expect(mocker.groupLog([123, 'abc'])).toBe(undefined);
+
+    console.groupCollapsed = undefined;
+    expect(mocker.groupLog([123, 'abc'])).toBe(undefined);
+    [console.groupCollapsed, console.groupEnd, console.log] = [log1, log2, log3];
+  });
+
+  it('MockItem.faker should be an instance of Faker.FakerStatic', () => {
+    const mockItem = new MockItem({
+      url: 'https://www.api.com/abc',
+      response: '<html>xxx</html>',
+      disable: 'true',
+    });
+    expect(mockItem.disable).toBe('yes');
+    expect(mockItem.faker).toBeTruthy();
+    expect(mockItem.faker.name).toBeTruthy();
+    expect(typeof mockItem.faker.name.findName()).toBe('string');
   });
 });